+++ src/stringfuncs.c
@@ -47,21 +47,16 @@ ULONG convertVersionToInt(STRPTR version)
   char buf[10], *tmp;
   uint32 versionNumber = 0;
 
-  IDOS->Printf("convertVersionToInt: %s\n", version);
-
   if (!strncmp(version, "v", 1))
     IUtility->Strlcpy(buf, version + 1, sizeof(buf));
 
-  IDOS->Printf("convertVersionToInt tmp=%s\n", buf);
-
   tmp = strtok(buf, ".");
-  IDOS->Printf("1> %s\n", tmp);
   versionNumber = versionNumber + (atoi(tmp) * 100);
+  
   tmp = strtok(NULL, ".");
-  IDOS->Printf("2> %s\n", tmp);
   versionNumber = versionNumber + (atoi(tmp) * 10);
+
   tmp = strtok(NULL, ".");
-  IDOS->Printf("3> %s\n", tmp);
   versionNumber = versionNumber + atoi(tmp);
 
   return versionNumber;